Moving catalogd controllers and webhook
Moving  catalogd/internal/controllers and catalogd/internal/webhooks
to internal/catalogd along with the Makefile changes
