+++ internal/catalogd/controllers/core/clustercatalog_controller.go
