+++ internal/catalogd/controllers/core/clustercatalog_controller_test.go
