+++ internal/catalogd/controllers/core/pull_secret_controller.go
