+++ internal/catalogd/controllers/core/pull_secret_controller_test.go
